+++ exp_by_day/10.30/exp11:i2v.sh
@@ -1,1 +1,2 @@
-bash ./scripts/train.sh -b 8 -c "2,3,4,5,6,7" -g 42 -w Exp11a -y i2v
+
+bash ./scripts/train.sh -b 8 -c "2,3,4,5,6,7" -g 42 -w Exp11a -y i2v -R wandb/10-31-191912_Exp11a/checkpoints/step200.ckpt